src/main/webapp/app/entities/eRestaurantCustomer/refunded/route/refunded-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IRefunded, Refunded } from '../refunded.model';
import { RefundedService } from '../service/refunded.service';

import { RefundedRoutingResolveService } from './refunded-routing-resolve.service';

describe('Refunded routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: RefundedRoutingResolveService;
  let service: RefundedService;
  let resultRefunded: IRefunded | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(RefundedRoutingResolveService);
    service = TestBed.inject(RefundedService);
    resultRefunded = undefined;
  });

  describe('resolve', () => {
    it('should return IRefunded returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRefunded = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultRefunded).toEqual({ id: 'ABC' });
    });

    it('should return new IRefunded if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRefunded = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultRefunded).toEqual(new Refunded());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Refunded })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultRefunded = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultRefunded).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
